move init logic to parent
